requirements.txt
Flask
dotenv
google-genai